Move parser.x into parser folder *
* Add comments on parser files *

* Up changelog. Fixes #13.
* Remove old TODOs
